tctl resource: convert lock handler
